tpm2: Restrict NV index datasize to MAX_NV_INDEX_SIZE (CID 1608875)
Resolve an issue reported by Coverity caused by the maximum value of
datasize (max. size of an NV index) that was allowed to be 0x10100
(17 bits) even though later on it tried to read an array of maximum size
expressed by 16 bits (Coverity complaint). However, the maximum value of
datasize could only ever have been MAX_NV_INDEX_SIZE, which is gated by
restrictions on the size of an NV index. Therefore, restrict the maximum
datasize of an NV index to MAX_NV_INDEX_SIZE (2048 bytes) since this is
the maximum size that an NV index can be defined for.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>