tpm2: In CheckLockedOut replace TPM_RC_RETRY with TPM_RC_SUCCESS
In CheckLockedOut replace the return code of TPM_RC_RETRY with
TPM_RC_SUCCESS since it does not seem to be necessary to run the TPM2
command again but the TPM2 can continue executing the current command.
If NVRAM wasn't available then the code in CheckLockedOut would return
with an error already.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>