tpm2: Make OaepDecode near constant-time (CVE-2026-6727)
This patch fixes CVE-2026-6727. Libtpms is typically not affected unless
it was configured with --disable-use-openssl-functions or if it was built
with '-DUSE_OPENSSL_FUNCTIONS_RSA=0'. Neither would be the 'typical' case.

OaepDecode exposes a potential timing side-channel. Fix the timing
side-channel of this function by accumulating a bad result in a variable
while still executing most of the code. This at least causes the test
case to not report the issue anymore.

Note: This issue only appears when using libtpms with
--disable-use-openssl-functions, which circumvents many OpenSSL functions
for RSA, ECDSA, and others, that likely handle constant-time issues better
than the TPM 2 reference code does.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>